- Added a project that includes the framework and test targets for each supported platform and also the Demo for iOS. - Made the schemes Shared for Carthage support. - Bumped the pod version. - Updated the Travis CI config.
