+++ .travis.yml
@@ -1,17 +1,31 @@
 # Reference: https://github.com/pinterest/PINRemoteImage/blob/master/.travis.yml
 #            https://github.com/Alamofire/Alamofire/blob/master/.travis.yml
 language: objective-c
-osx_image: xcode7.3
+osx_image: xcode8.3
 env:
   global:
   - LC_CTYPE=en_US.UTF-8
   - LANG=en_US.UTF-8
+  - IOS_FRAMEWORK_SCHEME="Loggerithm iOS"
+  - WATCHOS_FRAMEWORK_SCHEME="Loggerithm watchOS"
+  - TVOS_FRAMEWORK_SCHEME="Loggerithm tvOS"
+  - MACOS_FRAMEWORK_SCHEME="Loggerithm macOS"
+  - IOS_SDK=iphonesimulator10.3
+  - WATCHOS_SDK=watchsimulator3.2
+  - TVOS_SDK=appletvsimulator10.2
+  - MACOS_SDK=macosx10.12
+  - EXAMPLE_SCHEME="iOS Example"
   matrix:
-  - SCHEME=Demo SDK=iphonesimulator
-
+    - DESTINATION="OS=10.3,name=iPhone 7 Plus"     SCHEME="$IOS_FRAMEWORK_SCHEME"     SDK="$IOS_SDK"
+    - DESTINATION="OS=3.2,name=Apple Watch - 42mm" SCHEME="$WATCHOS_FRAMEWORK_SCHEME" SDK="$WATCHOS_SDK"
+    - DESTINATION="OS=10.2,name=Apple TV 1080p"    SCHEME="$TVOS_FRAMEWORK_SCHEME"    SDK="$TVOS_SDK"
+    - DESTINATION="arch=x86_64"                    SCHEME="$MACOS_FRAMEWORK_SCHEME"   SDK="$MACOS_SDK"
 before_install:
   - gem install cocoapods --no-rdoc --no-ri --no-document --quiet
   - gem install xcpretty --no-rdoc --no-ri --no-document --quiet
-
 script:
-  - xctool -project Demo/Demo.xcodeproj -scheme $SCHEME -sdk $SDK ONLY_ACTIVE_ARCH=NO
+  - set -o pipefail
+  - xcodebuild -version
+  - xcodebuild -showsdks
+script:
+  - xcodebuild -project Loggerithm.xcodeproj -scheme "$SCHEME" -sdk "$SDK" -destination "$DESTINATION" ONLY_ACTIVE_ARCH=NO | xcpretty;
