+++ Loggerithm.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:Demo.xcodeproj">
+      location = "self:Loggerithm.xcodeproj">
    </FileRef>
 </Workspace>
