+++ Source/Loggerithm.h
@@ -6,7 +6,7 @@
 //  Copyright (c) 2015 Honghao Zhang. All rights reserved.
 //
 
-#import <UIKit/UIKit.h>
+#import <Foundation/Foundation.h>
 
 //! Project version number for Loggerithm.
 FOUNDATION_EXPORT double LoggerithmVersionNumber;
